+++ framework/helpers/BaseJson.php
@@ -29,6 +29,17 @@ class BaseJson
      * @since 2.0.43
      */
     public static $prettyPrint = null;
+
+    /**
+     * @var bool Avoids objects with zero-indexed keys to be encoded as array
+     * `Json::encode((object)['test'])` will be encoded as an object not as an array. This matches the behaviour of `json_encode()`.
+     * Defaults to false to avoid any backwards compatibility issues.
+     * Enable for single purpose: `Json::$keepObjectType = true;`
+     * @see JsonResponseFormatter documentation to enable for all JSON responses
+     * @since 2.0.44
+     */
+    public static $keepObjectType = false;
+
     /**
      * List of JSON Error messages assigned to constant names for better handling of version differences.
      * @var array
@@ -159,6 +170,8 @@ class BaseJson
      */
     protected static function processData($data, &$expressions, $expPrefix)
     {
+        $revertToObject = false;
+
         if (is_object($data)) {
             if ($data instanceof JsExpression) {
                 $token = "!{[$expPrefix=" . count($expressions) . ']}!';
@@ -179,16 +192,28 @@ class BaseJson
                 $data = $data->toArray();
             } elseif ($data instanceof \SimpleXMLElement) {
                 $data = (array) $data;
+
+                // Avoid empty elements to be returned as array.
+                // Not breaking BC because empty array was always cast to stdClass before.
+                $revertToObject = true;
             } else {
+                /*
+                 * $data type is changed to array here and its elements will be processed further
+                 * We must cast $data back to object later to keep intended dictionary type in JSON.
+                 * Revert is only done when keepObjectType flag is provided to avoid breaking BC
+                 */
+                $revertToObject = static::$keepObjectType;
+
                 $result = [];
                 foreach ($data as $name => $value) {
                     $result[$name] = $value;
                 }
                 $data = $result;
-            }
 
-            if ($data === []) {
-                return new \stdClass();
+                // Avoid empty objects to be returned as array (would break BC without keepObjectType flag)
+                if ($data === []) {
+                    $revertToObject = true;
+                }
             }
         }
 
@@ -200,7 +225,7 @@ class BaseJson
             }
         }
 
-        return $data;
+        return $revertToObject ? (object) $data : $data;
     }
 
     /**